+++ typings/src/ui/index.ts
@@ -18,7 +18,6 @@ export * from './widgets';
 export * from './icon';
 export * from './colors';
 export * from './extensions';
-export * from './engagements';
 
 export enum IntegrationTypes {
   APP = 'App',
