+++ typings/src/ui/profile.ts
@@ -2,6 +2,7 @@ import {
   CeramicAccount,
   AkashaFollowConnection,
   AkashaProfile as ProfileData,
+  AkashaFollow,
 } from '../sdk/graphql-types-new';
 
 export type Profile = Omit<ProfileData, 'followers' | 'did'> & {
@@ -50,6 +51,17 @@ export enum StepStatus {
   GETTING_KEYS = 'gettingKeys',
 }
 
-export type EngagementType = 'followers' | 'following' | 'interests';
-
 export type ProfileImageType = 'avatar' | 'cover-image';
+
+export type AkashaFollowers = {
+  id: AkashaFollow['id'];
+  isFollowing: AkashaFollow['isFollowing'];
+  did?: Pick<CeramicAccount, 'akashaProfile'>;
+}[];
+
+export type AkashaFollowing = {
+  id: AkashaFollow['id'];
+  isFollowing: AkashaFollow['isFollowing'];
+  profile?: AkashaFollow['profile'];
+  did?: { id: string };
+}[];
